Limit concurrent Cassandra queries a bit.
Earlier there was no imposed limit and so we would fetch all partitions
for all attributes at the same time. For three attributes and one year
that would be around 1000 queries, which seems a bit much.

Probably too simplistic, but let's see if it helps against the worst
cases of stuck queries.