Fix a case where the panning/zooming goes crazy.
To reproduce, simply set a date or range in the dropdown, and when it's loaded
try to pan or zoom the data with the mouse. It jumps somewhere else in a very
confusing way.

This commit basically reverts the update to D3 v5, since that is where the bug
seems to have appeared.