Make cassandra consistency level configurable.
The consistency level determines how "strict" the cassandra driver is
about the data consistency when reading. By default, now it's content
to ask a single node, which means that we won't always get the very
latest data. I think this is OK in most cases, but we will see if
it causes problems. The setting is now configurable and therefore
easy to tweak.

The old setting of LOCAL_QUORUM causes problems because of how our
Cassandra and HDB++ systems are set up. If even a single node in the
cluster is down, the viewer would have problems.